+++ src/drivers/px4io/CMakeLists.txt
@@ -43,6 +43,7 @@ px4_add_module(
 		px4io_serial_f7.cpp
 	DEPENDS
 		circuit_breaker
+		mixer
 	)
 
 # include the px4io binary in ROMFS
