+++ src/drivers/snapdragon_pwm_out/CMakeLists.txt
@@ -38,4 +38,5 @@ px4_add_module(
 		snapdragon_pwm_out.cpp
 	DEPENDS
 		mixer
+		pwm_limit
 	)
