+++ src/drivers/tap_esc/CMakeLists.txt
@@ -39,5 +39,7 @@ px4_add_module(
 		tap_esc.cpp
 		tap_esc_common.cpp
 	DEPENDS
+		mixer
+		pwm_limit
 	)
