+++ src/lib/CMakeLists.txt
@@ -45,6 +45,7 @@ add_subdirectory(led)
 add_subdirectory(mathlib)
 add_subdirectory(mixer)
 add_subdirectory(perf)
+add_subdirectory(pwm_limit)
 add_subdirectory(rc)
 add_subdirectory(terrain_estimation)
 add_subdirectory(tunes)
