+++ src/lib/pwm_limit/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2018 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,63 +31,4 @@
 #
 ############################################################################
 
-px4_add_git_submodule(TARGET git_uavcan PATH "libuavcan")
-
-set(UAVCAN_USE_CPP03 ON CACHE BOOL "uavcan cpp03")
-set(UAVCAN_PLATFORM stm32 CACHE STRING "uavcan platform")
-
-if(NOT config_uavcan_num_ifaces)
-	message(FATAL_ERROR "config_uavcan_num_ifaces not set")
-endif()
-
-string(TOUPPER "${OS}" OS_UPPER)
-add_definitions(
-		-DUAVCAN_CPP_VERSION=UAVCAN_CPP03
-		-DUAVCAN_MEM_POOL_BLOCK_SIZE=48
-		-DUAVCAN_NO_ASSERTIONS
-		-DUAVCAN_PLATFORM=stm32
-		-DUAVCAN_STM32_${OS_UPPER}=1
-		-DUAVCAN_STM32_NUM_IFACES=${config_uavcan_num_ifaces}
-		-DUAVCAN_STM32_TIMER_NUMBER=5
-		)
-
-add_subdirectory(libuavcan EXCLUDE_FROM_ALL)
-add_dependencies(uavcan prebuild_targets)
-
-px4_add_module(
-	MODULE modules__uavcan
-	MAIN uavcan
-	STACK_MAIN 3200
-	STACK_MAX 1500
-	INCLUDES
-		${PX4_SOURCE_DIR}/mavlink/include/mavlink
-		libuavcan/libuavcan/include
-		libuavcan/libuavcan/include/dsdlc_generated
-		libuavcan/libuavcan_drivers/posix/include
-		libuavcan/libuavcan_drivers/stm32/driver/include
-	SRCS
-		# Main
-		uavcan_main.cpp
-		uavcan_servers.cpp
-		uavcan_params.c
-
-		# Actuators
-		actuators/esc.cpp
-		actuators/hardpoint.cpp
-
-		# Sensors
-		sensors/sensor_bridge.cpp
-		sensors/gnss.cpp
-		sensors/mag.cpp
-		sensors/baro.cpp
-
-	DEPENDS
-		mixer
-
-		git_uavcan
-
-		# within libuavcan
-		libuavcan_dsdlc
-		uavcan
-		uavcan_stm32_driver
-	)
+px4_add_library(pwm_limit pwm_limit.c)
