+++ src/lib/pwm_limit/pwm_limit.c
@@ -41,6 +41,7 @@
  */
 
 #include "pwm_limit.h"
+
 #include <math.h>
 #include <stdbool.h>
 #include <drivers/drv_hrt.h>
