+++ src/modules/vtol_att_control/CMakeLists.txt
@@ -42,5 +42,6 @@ px4_add_module(
 		tailsitter.cpp
 		standard.cpp
 	DEPENDS
+		pwm_limit
 	)
